rpc-grpc-protobuf-master/frontend/requirements.txt
flask==3.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0